pyproject.toml
[project]
name = "notification_service"
version = "0.1.0"
description = "Notification Service"
authors = [
    {name = "pavelvolkov",email = "swvolkovp@gmail.com"}
]
requires-python = ">=3.12.9"
dependencies = [
    "sqladmin (>=0.21.0,<0.22.0)",
    "faststream (>=0.6.1,<0.7.0)",
    "fastapi (>=0.119.0,<0.120.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.29.0,<1.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "uvicorn (>=0.30.0,<1.0.0)",
    "python-multipart (>=0.0.9,<1.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "google-api-python-client (>=2.185.0,<3.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "google-auth (>=2.41.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

